% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspoc.R
\name{CASPOC}
\alias{CASPOC}
\title{CASPOC}
\usage{
CASPOC(
  X,
  Y,
  ncomp = 1,
  numRepeats = 11,
  numFolds = 10,
  keepX_options = NULL,
  keepY_options = NULL,
  fixX = NULL,
  fixY = NULL,
  base_seed = 1,
  manual_folds = NULL
)
}
\arguments{
\item{X}{A matrix or vector}

\item{Y}{A matrix or vector}

\item{ncomp}{Number of components in the model}

\item{numRepeats}{Number of repeats for the cross-validation}

\item{numFolds}{Number of folds for the cross-validation}

\item{keepX_options}{Tune grid options for the keepX parameter - the number of included variables from X}

\item{keepY_options}{Tune grid options for the keepY parameter - the number of included variables from Y}

\item{fixX}{A vector of keepX values for each component; if you want to fix eg. keepX = 10 for comp1 but want to do a grid search on comp2. Leave as NULL for grid search on all components}

\item{fixY}{A vector of keepY values for each component; if you want to fix eg. keepY = 10 for comp1 but want to do a grid search on comp2. Leave as NULL for grid search on all components}

\item{base_seed}{Random seed for reproducibility. Use instead of 'set.seed()', since the function internally updates the seed between repeats.}

\item{manual_folds}{Manually supply folds. Should be a list of lists. Outer list should be of length numRepeats. Inner list should be of length numFolds and contain integer vectors supplying row indices for each fold.}
}
\value{
A list containing several elements:
\describe{
  \item{results_tune_df}{A data.frame with correlation results for each repeat and hyperparameter combination from the tuning folds}
  \item{results_test_df}{A data.frame with correlation results for each repeat and hyperparameter combination from the testing folds}
  \item{full_train_loadingsX}{A data.frame with X loadings for each variable in each fold, repeat and hyperparameter combination based on training folds}
  \item{full_train_loadingsY}{A data.frame with Y loadings for each variable in each fold, repeat and hyperparameter combination based on training folds}
  \item{full_tuneX}{A data.frame with all X component scores for each repeat and hyperparameter combination from the tuning folds}
  \item{full_tuneY}{A data.frame with all Y component scores for each repeat and hyperparameter combination from the tuning folds}
  \item{full_testX}{A data.frame with all X component scores for each repeat and hyperparameter combination from the testing folds}
  \item{full_testY}{A data.frame with all Y component scores for each repeat and hyperparameter combination from the testing folds}
  \item{folds}{Return the list of folds for CV}
  }
}
\description{
This is the main function in the caspoc package. It takes two matrices \(or vectors\), performs repeated K-fold cross-validation, and returns tune and test sets with all samples separately
}
\examples{
library(mixOmics)
data("breast.TCGA")
mydata_X <- breast.TCGA$data.train$mrna
mydata_Y <- breast.TCGA$data.train$protein
dim(mydata_X)
dim(mydata_Y)
#my_analysis <- CASPOC(mydata_X, mydata_Y, numRepeats = 5, numFolds = 10,
#   keepX_options = c(50, 100, 200), keepY_options = c(50, 100, 142),
#   ncomp = 1, base_seed = 42)
}
