% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caspoc.R
\name{deflate_sPLS_data}
\alias{deflate_sPLS_data}
\title{Deflation of data}
\usage{
deflate_sPLS_data(splsModel, ncomp, trainX, trainY, tuneX, tuneY, testX, testY)
}
\description{
This is an internal function used to deflate data across train, tune and test sets using training loadings. Only used if ncomp > 1
}
